requirements.txt
fastapi>=0.115,<1.0
uvicorn>=0.30,<1.0
jinja2>=3.1,<4.0
python-multipart>=0.0.9,<1.0
pytest>=8.0,<9.0
httpx>=0.27,<1.0
huggingface_hub>=0.36,<2.0
transformers>=4.57,<6.0
torch>=2.11,<3.0
numpy>=2.0,<3.0
pillow>=11.0,<12.0
kornia>=0.8,<1.0
safetensors>=0.6,<1.0
timm>=1.0,<2.0
torchvision>=0.26,<1.0
onnxruntime>=1.24,<2.0
